requirements.txt
streamlit==1.31.0
langchain==0.1.9
langchain-google-genai==1.0.1
langchain-community==0.0.24
google-generativeai>=0.4.1
PyPDF2==3.0.1
faiss-cpu==1.8.0
sentence-transformers>=5.0.0
python-dotenv==1.0.0
